chore: Switching mcore to upstream main
